Refactor, add comments, remove some console logging
