+++ rn/LogInView.js
@@ -3,6 +3,8 @@ import {Button, Text, TextInput, View} from 'react-native';
 import {styles} from './App';
 import {useAuth} from './AuthProvider';
 
+// This view has an input for email and password and logs in the user when the
+// "log in" button is pressed.
 export function LogInView() {
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
@@ -28,13 +30,13 @@ export function LogInView() {
       </View>
       <Button
         onPress={async () => {
-          console.log(`Login pressed with ${email} ${password}`);
+          console.log(
+            `Log in button pressed with email ${email} and password ${password}`,
+          );
           setError(null);
           try {
             await logIn(email, password);
-            console.log('Logged in');
           } catch (e) {
-            console.error('Caught it');
             setError(e);
           }
         }}
